Split FormDumper into individual implementations and simplify multiple constructs
