+++ scripts/commons/serviceportal/forms/FormDumper.groovy
@@ -56,80 +56,6 @@ class FormDumper {
   }
 
   /**
-   * Dump the content of a form as a HTML table formatted for the serviceportal.
-   *
-   * @param baseHeadingLevel the HTML-heading level for the top-most element (groups / accordions, etc)
-   * @return a String containing HTML code
-   */
-  String dumpFormAsHtmlTable(int baseHeadingLevel = 2) {
-    FormV1 formAndMapping = api.getForm(formContent.getFormId())
-    formAndMapping.setContent(formContent)
-
-    String result = ""
-
-    formAndMapping.groupInstances.each { instance ->
-      if (instance.isShown(formAndMapping)) {
-        // Determine the content of that instance first, so we can decide later if we want to
-        // show the (otherwise possibly empty) group
-        String groupContentRendered = ""
-        // check if instance is shown
-        instance.rows.each { FormRowV1 row ->
-          row.fields.each { FormFieldV1 field ->
-            // For rendering NPA fields hideDisabled must be false
-            if (shouldRenderField(field) && field.isShown(instance, formAndMapping)) {
-              groupContentRendered += "<tr>"
-
-              // Left column: The question
-              groupContentRendered += "<td>${field.label ?: ""}</td>"
-
-              // Right column: The answer
-              groupContentRendered += "<td>"
-
-
-              switch (field.type) {
-                case FieldTypeV1.GEO_MAP:
-                  // Special handling for GeoMap fields. Those fields are best represented as an image (rather than text)
-                  BinaryGeoMapContentV1 fieldContent = field.value as BinaryGeoMapContentV1
-                  groupContentRendered += generateEmbeddedImage(fieldContent.file)
-                  break
-                case FieldTypeV1.GDIK_MAP:
-                  // Special handling for GDIK map fields. Those fields are best represented as an image (rather than text)
-                  BinaryGDIKMapContentV1 fieldContent = field.value as BinaryGDIKMapContentV1
-                  for (file in fieldContent.files) {
-                    groupContentRendered += generateEmbeddedImage(file)
-                  }
-                  break
-                default:
-                  // Use a String representation for this field, as generated by renderFieldForUserOutput
-                  // Escape the result of renderFieldForUserOutput as it might contain XSS content
-                  groupContentRendered += org.apache.commons.text.StringEscapeUtils.escapeHtml4(renderFieldForUserOutput(field))
-              }
-
-              groupContentRendered += "</td>"
-
-              groupContentRendered += "</tr>"
-            }
-          }
-        }
-
-        // Now, build the instance
-        if (!groupContentRendered.isEmpty()) {
-          result += "<h${baseHeadingLevel}>${instance.title}</h${baseHeadingLevel}>"
-          // General headings for the instance
-          result += "<table class=\"summary-form\">"
-          result += "<thead><tr><th>Feld</th><th>Ihre Eingabe</th></tr></thead>"
-          result += "<tbody>"
-          result += groupContentRendered
-          result += "</tbody>"
-          result += "</table>"
-        }
-      }
-    }
-
-    return result
-  }
-
-  /**
    * Dump the content of a form as a human readable text
    *
    * @param printGroupHeadings set to true if groups in the form should have headings
@@ -413,112 +339,6 @@ class FormDumper {
     return new JsonBuilder(fieldsAndValues).toPrettyString()
   }
 
-  /**
-   * Get a String representation useful for displaying it back to the user.
-   * E.g. Checkboxes will contain the label the user selected, dates are formatted with a German
-   * date format and no time component, etc.
-   *
-   * @param field The field to render
-   * @return The String representation
-   */
-  @SuppressWarnings('GrDeprecatedAPIUsage')
-  // We need to support deprecated form field types as they might still be in use by older forms
-  private String renderFieldForUserOutput(FormFieldV1 field) {
-    Object value = field.value
-    if (value instanceof VerifiedFormFieldValueV1)
-    {
-      value = ((VerifiedFormFieldValueV1) value).value
-    }
-    if (value == null || value.toString().isAllWhitespace()) {
-      return "[Keine Eingabe]"
-    } else {
-      //noinspection GroovyFallthrough - those fall-throughs are on purpose.
-      switch (field.type) {
-        case FieldTypeV1.STRING:
-          // fall through
-        case FieldTypeV1.STRING_AJAX_AUTOCOMPLETE:
-          // fall through
-        case FieldTypeV1.KFZ_KENNZEICHEN:
-          // fall through
-        case FieldTypeV1.TEXTAREA:
-          return value
-          break
-        case FieldTypeV1.FILE:
-          return "Datei: \"${(value as BinaryContentV1).uploadedFilename}\""
-          break
-        case FieldTypeV1.MULTIPLE_FILE:
-          return getFilenamesFromMultipleUpload(field)
-          break
-        case FieldTypeV1.BOOLEAN:
-          // fall though
-        case FieldTypeV1.SINGLE_CHECKBOX:
-          return value ? "Ja" : "Nein"
-          break
-        case FieldTypeV1.CHECKBOX:
-          return generateCommaSeparatedListOfPossibleValueLabel(value as ArrayList<String>, field.possibleValues)
-          break
-        case FieldTypeV1.RADIO_BUTTONS:
-          return findLabelForPossibleValue(field.possibleValues, value as String)
-          break
-        case FieldTypeV1.DROPDOWN_SINGLE_SELECT:
-          // fall-through
-        case FieldTypeV1.DROPDOWN_SINGLE_SELECT_AJAX:
-          return findLabelForPossibleValue(field.possibleValues, value as String)
-          break
-        case FieldTypeV1.DROPDOWN_MULTIPLE_SELECT:
-          return generateCommaSeparatedListOfPossibleValueLabel(value as ArrayList<String>, field.possibleValues)
-          break
-        case FieldTypeV1.TWO_LIST_SELECT:
-          return generateCommaSeparatedListOfPossibleValueLabel(value as ArrayList<String>, field.possibleValues)
-          break
-        case FieldTypeV1.DATE:
-          return new SimpleDateFormat("dd.MM.yyyy").format(value as Date)
-          break
-        case FieldTypeV1.TIME:
-          return new SimpleDateFormat("HH:mm").format(value as Date)
-          break
-        case FieldTypeV1.EURO_BETRAG:
-          return (value as BigDecimal).toString() + " €"
-          break
-        case FieldTypeV1.SUBMITTED_WITH_NPA_INFO:
-          return value ? "Sie waren mit dem neuem Personalausweis angemeldet" : "Sie waren NICHT mit dem neuem Personalausweis angemeldet"
-          break
-        case FieldTypeV1.TEXT:
-          // fall through
-        case FieldTypeV1.H2:
-          // fall through
-        case FieldTypeV1.H1:
-          // fall through
-        case FieldTypeV1.PLACEHOLDER:
-          // fall through
-        case FieldTypeV1.DOWNLOAD:
-          // fall through
-        case FieldTypeV1.VIDEO:
-          // fall through
-        case FieldTypeV1.IMAGE:
-          // User can't input data in this field type. So nothing is rendered.
-          return ""
-          break
-        case FieldTypeV1.GEO_MAP:
-          // Its difficult to represent as GeoMap field as a String, so we just output the attributes we get from the API.
-          // Note that the methods calling renderFieldForUserOutput (like renderAsHtmlTable()) might have overridden and
-          // more suitable behaviour than representing the GeoMap field as a String.
-          BinaryGeoMapContentV1 binaryGeoMapContentValue = field.value as BinaryGeoMapContentV1
-          return "Nutzereingaben: '${binaryGeoMapContentValue.json}', Auswahl von Elementen auf der Karte: '${binaryGeoMapContentValue.selectionJson}'"
-        case FieldTypeV1.GDIK_MAP:
-          // Similar to GEO_MAP, its difficult to represent GDIK_MAP as a String.
-          BinaryGDIKMapContentV1 binaryGDIKMapContentValue = field.value as BinaryGDIKMapContentV1
-          return "Nutzereingaben: '${binaryGDIKMapContentValue.json}', Auswahl von Elementen auf der Karte: '${binaryGDIKMapContentValue.selectionJson}'"
-        default:
-          api.logger.warn("FormDumper.renderFieldForUserOutput does not know how to display this field '${field.type}' (${field.type.class.name}), " + "so it defaults to toString().")
-          return value.toString()
-          break
-      }
-    }
-
-    throw new RuntimeException("Unexpected field type '${field.type}'. " + "The FormDumper class does not know how to render that.")
-  }
-
   private String renderFieldForXmlOutput(FormFieldV1 field) {
     def value = field.value
 
@@ -568,89 +388,9 @@ class FormDumper {
     }
   }
 
-  private static String generateCommaSeparatedListOfPossibleValueLabel(List<String> values, List<PossibleValueV1> pvList) {
-    String result = ""
-    boolean isFirst = true
-    values.each { technicalName ->
-      if (!isFirst) {
-        // Add separator
-        result += ", "
-      }
-      result += findLabelForPossibleValue(pvList, technicalName)
-      isFirst = false
-    }
-    return result
-  }
-
-  private static String findLabelForPossibleValue(List<PossibleValueV1> pvList, String value) {
-    PossibleValueV1 pv = pvList.find { it.value == value }
-
-    if (pv != null) {
-      return pv.label
-    } else {
-      throw new IllegalArgumentException("Could not find possible value '$value' in PossibleValueList '$pvList'")
-    }
-  }
-
-  private static String getFilenamesFromMultipleUpload(FormFieldV1 field) {
-    String result = ""
-    field.value.eachWithIndex { it, idx ->
-      if (idx == (field.value as List<BinaryContentV1>).size() - 1) {
-        result += "\"${(it as BinaryContentV1).uploadedFilename}\""
-      } else {
-        result += "\"${(it as BinaryContentV1).uploadedFilename}\", "
-      }
-    }
-    return result.strip()
-  }
-
-  /**
-   * Determines if a field should be rendered or not. Fields will not be rendered, if they are of a type where users can
-   * not enter any input (like headings, download buttons, etc.) or if additional logic from the
-   * additionalLogicToHideContent parameter should be applied.
-   *
-   * @param field the field to check
-   *
-   * @return true, if the field should be shown
-   */
-  private boolean shouldRenderField(FormFieldV1 field) {
-    // Don't show TEXT-fields the user can't input into
-    if (field.type == FieldTypeV1.TEXT ||
-            field.type == FieldTypeV1.H2 ||
-            field.type == FieldTypeV1.H1 ||
-            field.type == FieldTypeV1.PLACEHOLDER ||
-            field.type == FieldTypeV1.DOWNLOAD ||
-            field.type == FieldTypeV1.VIDEO ||
-            field.type == FieldTypeV1.IMAGE ||
-            field.type == FieldTypeV1.HINTBOX) {
-      return false
-    }
-
-    // User might have overridden the 'additionalLogicToHideContent' to hide some additional fields.
-    // That's why we evaluate that closure now.
-    if (additionalLogicToHideContent.call(field) == true) {
-      return false
-    }
-
-    // **DO** show all other fields
-    return true
-  }
-
   private static String escapeForCsv(String stringToEscape) {
     // Add surrounding quotes and escape quotes that might appear in the string
     return "\"" + stringToEscape.replace("\"", "\"\"") + "\""
-  }
-
-  /**
-   * Generates an HTML tag containing an image file as base64 encoded data
-   *
-   * @param imageFile file containing the image data
-   *
-   * @return
-   */
-  private static String generateEmbeddedImage(BinaryContentV1 imageFile) {
-    String base64OfImage = Base64.getEncoder().encodeToString(imageFile.data)
-    return "<img src='data:image/jpeg;base64,${base64OfImage}' width='100%'>"
   }
 
   /**
